Cleaned up code
Focused on the ORM style of modeling the database and created basic CRUD functions with that style.

+++ main.py
@@ -1,10 +1,11 @@
-from sqlalchemy import create_engine, text, insert
-from sqlalchemy.orm import DeclarativeBase
-from sqlalchemy.orm import Mapped
+from sqlalchemy import String, Uuid
+from sqlalchemy import create_engine, select
+from uuid import UUID, uuid4
+from sqlalchemy.orm import DeclarativeBase, Mapped, Session
 from sqlalchemy.orm import mapped_column
-from sqlalchemy.orm import Session
-from sqlalchemy import String, select, Uuid
-import uuid
+
+# The engine declares where the database is as well as the credentials to use it
+engine = create_engine("postgresql://postgres:Coder119!@localhost:5435/userdb", echo=False)
 
 class Base(DeclarativeBase):
     pass
@@ -12,7 +13,7 @@ class Base(DeclarativeBase):
 class Users(Base):
     __tablename__ = 'users'
 
-    id: Mapped[str] = mapped_column(Uuid, primary_key=True)
+    id: Mapped[UUID] = mapped_column(Uuid(as_uuid=True), primary_key=True, default=uuid4)
     name: Mapped[str] = mapped_column(String(100), nullable=False)
     email: Mapped[str] = mapped_column(String(50), nullable=False, unique=True)
 
@@ -20,67 +21,69 @@ class Users(Base):
     def __repr__(self) -> str:
         return f"User(id = {self.id!r}), name = {self.name!r}, email = {self.email!r}"
     
-def add_user(engine, name: str, email: str):
-    session = Session(engine)
-
-    new_user = Users(id=uuid.uuid4(), name=name, email=email)
-
-    session.add(new_user)
-    session.commit()
-    session.close()
-
-def get_all(engine):
-    session = Session(engine)
-    stmt = select(Users)
-
-    return session.scalars(stmt)
-
-def get_by_id(engine, id: str):
-    session = Session(engine)
-
-    return session.get(Users,id)
-
-def update_user(engine, id: str, name: str = '', email: str = ''):
-    session = Session(engine)
-
-    stmt = session.execute(select(Users).filter_by(id=id)).scalar_one()
-
-    if name != '':
-        stmt.name = name
-
-    if email != '':
-        stmt.email = email
-    
-    session.commit()
-    session.close()
-
-def delete_user(engine, id: str):
-    session = Session(engine)
-
-    user = session.get(Users, id)
-
-    session.delete(user)
-
-    session.commit()
-    session.close()
-
-
-engine = create_engine("postgresql://postgres:Coder119!@localhost:5435/userdb", echo=False)
+    def __str__(self) -> str:
+        return f"User(id = {self.id!r}), name = {self.name!r}, email = {self.email!r}"
     
 Base.metadata.create_all(engine)
 
-# session = Session(engine)
 
-# add_user(engine, 'Ryan456', 'notarealemail@ymail.com')
+def get_all(session: Session):
+    stmt = select(Users)
+    return session.scalars(stmt).all()
 
-session = Session(engine)
+def get_by_id(session: Session, uuid: UUID):
+    return session.get(Users, uuid)
 
-get_user = get_by_id(engine, '7b6745e2-19ee-4984-b139-927bdff5e836')
+def get_by_email(session: Session, email: str):
+    return session.scalars(select(Users).where(Users.email == email)).one_or_none()
 
-print(get_user)
+def add_user(session: Session, name: str, email: str):
+    new_user = Users(name=name, email=email)
+    session.add(new_user)
+    session.commit()
 
-delete_user(engine, '7b6745e2-19ee-4984-b139-927bdff5e836')
+def update_user(session: Session, id: UUID, name: str | None = None, email: str | None = None):
+    u = session.get(Users, id)
 
-get_user = get_by_id(engine, '7b6745e2-19ee-4984-b139-927bdff5e836')
+    if u is not None:
+        if name is not None:
+            u.name = name
 
-print(get_user)
+        if email is not None:
+            u.email = email
+
+        session.commit()
+
+    else:
+        print(f"No user with id {id} found.")
+
+def delete_user(session: Session, id: UUID):
+    user = session.get(Users, id)
+
+    if user is not None:
+        session.delete(user)
+        session.commit()
+
+    else:
+        print(f"No user with id {id} found.")
+
+# with Session(engine) as session:
+
+#     new_user = "Aaron"
+#     email = "aaron@email.com"
+
+#     # add_user(session, new_user, email)
+
+#     aaron_id = '77b88e48-6118-477a-8fe3-35f7b5803afc'
+
+#     print(get_by_id(session, aaron_id))
+
+#     new_email = "aaron@microsoft.com"
+
+#     update_user(session, aaron_id, email=new_email)
+
+#     print(get_by_id(session, aaron_id))
+
+#     delete_user(session, aaron_id)
+
+#     print(get_by_id(session, aaron_id))
